Added a firewall_manager parameter
To define what module to use for managing the firewall, added a new
param called firewall_manager. This is set to either "iptables" or
"firewall" depending on whether the user is running CentOS 6 or CentOS
7. Other operating systems throw an error.

The new parameter is used in the main class when determining what to do
about setting firewall rules.