Patch to fix the opsview ordering issues
There is a known issue with the opsview-agent. When the package installs
from the repository it automatically starts the service. Puppet then
tries to start it but its already running so the service isn't updated
with the latest configs. This adds a restart as start to mitigate this
issue. opsview-agent is a bad package and this was the simplest
solution.
Also, service notifications were updated

Added ristos suggestions for notification