Use service tenant name not id in capacity check (#1)
Fix an issue that was preventing check_openstack from working when the
service tenant id was something other than "service".

When filtering for the public provider network it should not be searched
for based on id. The id can change based on environment and can't be
assumed to always be the same as was the case previously. With this
patch we make a slightly less bad assumption that the service tenant is
always named "service".