Usage stats speed improvement + add util stats (#5)
Previously the usage stats script was unnecessarily making two heavy API
calls multiple times: a query of all VMs and a query of all users. Run
each of these queries only once and use the rest of the functions to
only do filtering. This improves the running time of the script from a
little over 30 seconds to a little under 10 seconds.

Add a new function for getting statistics about the total utilization of
all compute nodes. Using this function, make the script return three
additional values in the tuple that it outputs: used memory on all
hypervisors, total memory of all hypervisors and a utilization
percentage calculated from these two values.

Also, the optparse module is deprecated in Python 2.7. Use the newer
argparse instead for parsing command line arguments.