Add RabbitMQ monitoring scripts
Add monitoring scripts for RabbitMQ. The scripts are from here:
https://github.com/nagios-plugins-rabbitmq/nagios-plugins-rabbitmq

These are written in Perl, so the user of these scripts will need to
install some Perl dependencies. On CentOS, these would be (via yum):

 - perl-URI
 - perl-JSON
 - perl-LWP-UserAgent-Determined
 - perl-Monitoring-Plugin

Installing those will obviously also pull in Perl if it is not already
installed. Installation is left to the module user - the scripts are
added here only for distribution purposes.