Reduce output of these two nagios checks to under 1024 characters (#41)
* Option to only print windows aggregate

Somewhere there is a limit that limits check_nrpe output to 1024
characters. This is a workaround. One can now configure a second
"check_openstack.py capacity -k" which only shows windows aggregate
stuff (as well as all the other details from OSCapacity).

* Rename "num_vms_by_flavors" to "vms"

To get the characters from 1364 to 898 in prod

 - #CCCP-2193