Add smaller capcity checks to check_openstack.py
This tries to fix an issue where our 'capacity' check returns a too
long string to our loging.

This would make it possible to have different aggregate checks for
cpus, ram, and other. The only adds three checks that return same
but less data as the 'capacity' check. The following checks are:

capacityram, capacitycpus, capacitysimple

CCCP-2753