Add smaller capcity checks to check_openstack.py (#51)
* Add smaller capacity checks to check_openstack.py

This tries to fix an issue where our 'capacity' check returns a too
long string to our logging.

This would make it possible to have different aggregate checks for
cpus, ram, and other. The only adds three checks that return same
but less data as the 'capacity' check. The following checks are:

capacityram, capacitycpus, capacitynetwork

CCCP-2753