Fixed check_nrpe regex bug
The current version of the check_nrpe check unmanaged(maintenance mode)
verification check was not working as the regex was matching for
alphanumeric strings where the hostname was in crm status output.
Because of the use of FQDN with dots and hyphens it caused the regex to
fail.

Switching from \w+  (a-z, A-Z, 0-9, _) ( + being one or more )
to \S+   (any non-white space character) ( + being one or more )

meant that the full hostname was captured by the regex as intended.