This script is to check difference between hardware clock and system clock
Some time, for various reason hardware clock runs behind the
system time. This may create problem for various services running on
system (for example puppetmaster may create a certificate which will
be activated in future, which can fail system to be cert signed).
We need to get notified when hwclock is running behind the system time.

This script will notify and send exit code 1 (warning for nagios) when
hwclock is in past more than 15minutes. Also exit code 0 (ok for nagios)
when difference between hwclock and system clock is less than 10s.

- CCCP-2845