Script to check difference between hardware clock and system clock (#54)
* This script is to check difference between hardware clock and system clock

Some time, for various reason hardware clock runs behind the
system time. This may create problem for various services running on
system (for example puppetmaster may create a certificate which will
be activated in future, which can fail system to be cert signed).
We need to get notified when hwclock is running out of sync with the system time.