Add script to check Windows instances
* Add check for Windows instances

We would benefit from the possibility to monitor that Windows instances
are running inside their designated aggregates. The check in this commit
tried to achieve this.

* Define identity API version to v3

* Fix for utf8 server names in python 2

Python 2 prints UTF-8 strings as follows: u'this is my string'
The current sed filtering does not cover this case. This commit tries to
enhance sed filtering accordingly.

* Fix typo in OK message

* Fix according to ShellCheck prescriptions

* Simplify command to get list of hosts in aggregate