Add date check
If the key generation process does not log any message to the log file,
this could potentially lead to a false positive if the last line in the
log file is about a succesful run. In other words, the Nagios check
would keep saying that everything is fine, while the keys aren't being
generated in the first place.
This commit adds a check on the date of the latest logged message: if
the date is different from today's date, an error is raised.