Add monitoring for Fernet keys generation process (#81)
* Add Fernet key generation process check

The generation process of the Fernet keys is currently not monitored.
This commit adds a check that can be used by Nagios to check the status
of the latest run of the key generation process.

* Simplify check_fernet_keys.py

subcommand can be a bit dangerous and the .run call is not supported in
python 2.7. This commit changes the check so that the subcommand module
is not needed any longer.

* Fix syntax error

* Add date check

If the key generation process does not log any message to the log file,
this could potentially lead to a false positive if the last line in the
log file is about a succesful run. In other words, the Nagios check
would keep saying that everything is fine, while the keys aren't being
generated in the first place.
This commit adds a check on the date of the latest logged message: if
the date is different from today's date, an error is raised.